Add option to print out src coverage with functions (#14)
* move logger creation to method

* fix print functions

* first try for printing source

* statements are hard to match against blocks

* src output with color working

* add print src flag

* tests and linting

+++ pkg/coverage/parser/goparser/statements/statements.go
@@ -15,11 +15,11 @@
 package statements
 
 type Statement struct {
-	StartOffset   int
-	StartLine     int
-	StartCol      int
-	EndOffset     int
-	EndLine       int
-	EndCol        int
-	ExecutedCount int
+	StartOffset   int64
+	StartLine     int64
+	StartCol      int64
+	EndOffset     int64
+	EndLine       int64
+	EndCol        int64
+	ExecutedCount int64
 }
